+++ rastervision/new_version/learner/metrics.py
@@ -1,0 +1,39 @@
+import torch
+
+
+def compute_conf_mat(out, y, num_labels):
+    labels = torch.arange(0, num_labels).to(out.device)
+    return ((out == labels[:, None]) & (y == labels[:, None, None])).sum(
+        dim=2, dtype=torch.float32)
+
+
+def compute_conf_mat_metrics(conf_mat, label_names, eps=1e-6):
+    # eps is to avoid dividing by zero.
+    eps = torch.tensor(eps)
+    conf_mat = conf_mat.cpu()
+    gt_count = conf_mat.sum(dim=1)
+    pred_count = conf_mat.sum(dim=0)
+    total = conf_mat.sum()
+    true_pos = torch.diag(conf_mat)
+    precision = true_pos / torch.max(pred_count, eps)
+    recall = true_pos / torch.max(gt_count, eps)
+    f1 = (2 * precision * recall) / torch.max(precision + recall, eps)
+
+    weights = gt_count / total
+    weighted_precision = (weights * precision).sum()
+    weighted_recall = (weights * recall).sum()
+    weighted_f1 = ((2 * weighted_precision * weighted_recall) / torch.max(
+        weighted_precision + weighted_recall, eps))
+
+    metrics = {
+        'avg_precision': weighted_precision.item(),
+        'avg_recall': weighted_recall.item(),
+        'avg_f1': weighted_f1.item()
+    }
+    for ind, label in enumerate(label_names):
+        metrics.update({
+            '{}_precision'.format(label): precision[ind].item(),
+            '{}_recall'.format(label): recall[ind].item(),
+            '{}_f1'.format(label): f1[ind].item(),
+        })
+    return metrics