+++ rastervision/v2/rv/data/label_source/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+
+from rastervision.v2.rv.data.label_source.label_source import *
+from rastervision.v2.rv.data.label_source.label_source_config import *
+from rastervision.v2.rv.data.label_source.chip_classification_label_source import *
+from rastervision.v2.rv.data.label_source.chip_classification_label_source_config import *