+++ rastervision/v2/rv/evaluation/__init__.py
@@ -1,0 +1,12 @@
+# flake8: noqa
+
+from rastervision.v2.rv.evaluation.evaluation_item import *
+from rastervision.v2.rv.evaluation.class_evaluation_item import *
+from rastervision.v2.rv.evaluation.evaluator import *
+from rastervision.v2.rv.evaluation.evaluator_config import *
+from rastervision.v2.rv.evaluation.classification_evaluation import *
+from rastervision.v2.rv.evaluation.chip_classification_evaluation import *
+from rastervision.v2.rv.evaluation.classification_evaluator import *
+from rastervision.v2.rv.evaluation.classification_evaluator_config import *
+from rastervision.v2.rv.evaluation.chip_classification_evaluator import *
+from rastervision.v2.rv.evaluation.chip_classification_evaluator_config import *